Renamed utils file
- Remove licencing info
- Addressed PR change requests (https://github.com/ska-sa/astrokat/pull/63)

JIRA: [MT-828](https://skaafrica.atlassian.net/browse/MT-828)
