+++ astrokat/test/testutils.py
@@ -1,11 +1,3 @@
-###############################################################################
-# SKA South Africa (http://ska.ac.za/)                                        #
-# Author: cam@ska.ac.za                                                       #
-# Copyright @ 2019 SKA SA. All rights reserved.                               #
-#                                                                             #
-# THIS SOFTWARE MAY NOT BE COPIED OR DISTRIBUTED IN ANY FORM WITHOUT THE      #
-# WRITTEN PERMISSION OF SKA SA.                                               #
-###############################################################################
 import logging
 import os
 from six.moves import StringIO
@@ -17,8 +9,15 @@ __all__ = ['yaml_path', 'LoggedTelescope']
 
 
 def yaml_path(file_path):
+    """Convenience method for finding the yaml's file absolute path.
+
+    Args:
+        file_path (str): YAML file path
+
+    Returns:
+        str: YAML file absolute path
+    """
     tests_path = os.path.abspath(os.path.dirname(__file__))
-    """Convenience method for finding the yaml's file absolute path."""
     yaml_file = os.path.abspath(os.path.join(tests_path, file_path))
     assert os.path.isfile(yaml_file)
     return yaml_file
@@ -26,13 +25,23 @@ def yaml_path(file_path):
 
 class LoggedTelescope(observe_main.Telescope):
 
-    string_stream = StringIO()
+    """
+    Class to be used as class decorator for test case.
+    The body of the test case class is patched with a new object.
+    When the class exits the patch is undone.
+
+    Attributes:
+        user_logger_stream (_io.StringIO): Text I/O implementation using an in-memory
+            buffer.
+    """
+
+    user_logger_stream = StringIO()
 
     def __init__(self, *args, **kwargs):
         super(LoggedTelescope, self).__init__(*args, **kwargs)
         # Add log handler AFTER init, as the user_logger is
         # replaced during init
-        out_hdlr = logging.StreamHandler(self.string_stream)
+        out_hdlr = logging.StreamHandler(self.user_logger_stream)
         formatter = logging.Formatter('%(asctime)s - %(message)s')
         formatter.formatTime = simulate.sim_time
         out_hdlr.setFormatter(formatter)
